+++ examples/discovery-search-app/cypress/support/index.ts
@@ -13,8 +13,9 @@
 // https://on.cypress.io/configuration
 // ***********************************************************
 
+//TODO: once the typescript + cypress PR is merged, we can change commands.js -> commands.ts
 // Import commands.js using ES2015 syntax:
-// import './commands.ts';
+import './commands.js';
 
 // import the default settings for cy.server()
 import './serverDefaults.ts';