+++ packages/discovery-react-components/src/components/DocumentPreview/__fixtures__/passages.ts
@@ -1,14 +1,14 @@
 export default {
   single: {
     passage_text: '5.21 Miscellaneous Costs',
-    start_offset: 39574,
-    end_offset: 39598,
+    start_offset: 39611,
+    end_offset: 39635,
     field: 'text'
   },
   multiline: {
     passage_text: `32.7 Buyer’s and Customer's Regulatory Authorities shall have the benefit of any rights of audit and access to information and documentation provided for in this Agreement to the extent that they relate to the exercise of the Regulatory Authorities' legal rights and/or responsibilities.`,
-    start_offset: 138570,
-    end_offset: 138857,
+    start_offset: 138812,
+    end_offset: 139099,
     field: 'text'
   }
 };