+++ packages/discovery-react-components/CHANGELOG.md
@@ -3,6 +3,77 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0-beta.20](https://github.com/watson-developer-cloud/discovery-components/compare/@ibm-watson/discovery-react-components@1.1.0-beta.19...@ibm-watson/discovery-react-components@1.1.0-beta.20) (2020-01-22)
+
+
+### Bug Fixes
+
+* calculate DOM offset first, followed by TextNode ([#57](https://github.com/watson-developer-cloud/discovery-components/issues/57)) ([4215c2d](https://github.com/watson-developer-cloud/discovery-components/commit/4215c2d))
+
+
+### Features
+
+* remove pdfFallback component ([#66](https://github.com/watson-developer-cloud/discovery-components/issues/66)) ([2d56124](https://github.com/watson-developer-cloud/discovery-components/commit/2d56124))
+
+
+
+
+
+# [1.1.0-beta.19](https://github.com/watson-developer-cloud/discovery-components/compare/@ibm-watson/discovery-react-components@1.1.0-beta.18...@ibm-watson/discovery-react-components@1.1.0-beta.19) (2020-01-22)
+
+**Note:** Version bump only for package @ibm-watson/discovery-react-components
+
+
+
+
+
+# [1.1.0-beta.18](https://github.com/watson-developer-cloud/discovery-components/compare/@ibm-watson/discovery-react-components@1.1.0-beta.17...@ibm-watson/discovery-react-components@1.1.0-beta.18) (2020-01-22)
+
+**Note:** Version bump only for package @ibm-watson/discovery-react-components
+
+
+
+
+
+# [1.1.0-beta.17](https://github.com/watson-developer-cloud/discovery-components/compare/@ibm-watson/discovery-react-components@1.1.0-beta.16...@ibm-watson/discovery-react-components@1.1.0-beta.17) (2020-01-21)
+
+**Note:** Version bump only for package @ibm-watson/discovery-react-components
+
+
+
+
+
+# [1.1.0-beta.16](https://github.com/watson-developer-cloud/discovery-components/compare/@ibm-watson/discovery-react-components@1.0.7...@ibm-watson/discovery-react-components@1.1.0-beta.16) (2020-01-21)
+
+
+### Bug Fixes
+
+* cidocument test issue ([#21](https://github.com/watson-developer-cloud/discovery-components/issues/21)) ([257b2e8](https://github.com/watson-developer-cloud/discovery-components/commit/257b2e8))
+* cidocument test quickfix ([#20](https://github.com/watson-developer-cloud/discovery-components/issues/20)) ([3586fc7](https://github.com/watson-developer-cloud/discovery-components/commit/3586fc7))
+* error in document preview with split docs ([#29](https://github.com/watson-developer-cloud/discovery-components/issues/29)) ([ed47956](https://github.com/watson-developer-cloud/discovery-components/commit/ed47956))
+* Remove node tilde ([#40](https://github.com/watson-developer-cloud/discovery-components/issues/40)) ([303ccb1](https://github.com/watson-developer-cloud/discovery-components/commit/303ccb1))
+* use react-scripts to run with code-coverage ([#61](https://github.com/watson-developer-cloud/discovery-components/issues/61)) ([1f78203](https://github.com/watson-developer-cloud/discovery-components/commit/1f78203))
+
+
+### Features
+
+* add i18n for ResultsPagination component ([#58](https://github.com/watson-developer-cloud/discovery-components/issues/58)) ([1bacb7b](https://github.com/watson-developer-cloud/discovery-components/commit/1bacb7b))
+* structured query add and remove group rules ([#42](https://github.com/watson-developer-cloud/discovery-components/issues/42)) ([8f9ebb1](https://github.com/watson-developer-cloud/discovery-components/commit/8f9ebb1))
+* structured query add and remove rules ([#27](https://github.com/watson-developer-cloud/discovery-components/issues/27)) ([eab1d3b](https://github.com/watson-developer-cloud/discovery-components/commit/eab1d3b))
+* structured query basic ([#22](https://github.com/watson-developer-cloud/discovery-components/issues/22)) ([b992f63](https://github.com/watson-developer-cloud/discovery-components/commit/b992f63))
+* upgrade storybook to 5.3 ([#25](https://github.com/watson-developer-cloud/discovery-components/issues/25)) ([93db599](https://github.com/watson-developer-cloud/discovery-components/commit/93db599))
+
+
+
+
+
+## [1.0.7](https://github.com/watson-developer-cloud/discovery-components/compare/@ibm-watson/discovery-react-components@1.0.7-rc.1...@ibm-watson/discovery-react-components@1.0.7) (2020-01-21)
+
+**Note:** Version bump only for package @ibm-watson/discovery-react-components
+
+
+
+
 # [1.1.0-beta.15](https://github.com/watson-developer-cloud/discovery-components/compare/@ibm-watson/discovery-react-components@1.1.0-beta.14...@ibm-watson/discovery-react-components@1.1.0-beta.15) (2020-01-20)
 
 
@@ -16,11 +87,31 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [1.1.0-beta.14](https://github.com/watson-developer-cloud/discovery-components/compare/@ibm-watson/discovery-react-components@1.1.0-beta.13...@ibm-watson/discovery-react-components@1.1.0-beta.14) (2020-01-16)
 
-
 ### Features
 
 * add i18n for ResultsPagination component ([#58](https://github.com/watson-developer-cloud/discovery-components/issues/58)) ([1bacb7b](https://github.com/watson-developer-cloud/discovery-components/commit/1bacb7b))
 
+
+
+
+## [1.0.7-rc.1](https://github.com/watson-developer-cloud/discovery-components/compare/@ibm-watson/discovery-react-components@1.0.7-rc.0...@ibm-watson/discovery-react-components@1.0.7-rc.1) (2020-01-15)
+
+### Features
+
+* standalone HTML view with table highlighting ([#48](https://github.com/watson-developer-cloud/discovery-components/issues/48)) ([d8b1207](https://github.com/watson-developer-cloud/discovery-components/commit/d8b1207))
+* allow custom document text field name ([#55](https://github.com/watson-developer-cloud/discovery-components/issues/55)) ([c24363f](https://github.com/watson-developer-cloud/discovery-components/commit/c24363f))
+
+
+
+
+## [1.0.7-rc.0](https://github.com/watson-developer-cloud/discovery-components/compare/@ibm-watson/discovery-react-components@1.0.6...@ibm-watson/discovery-react-components@1.0.7-rc.0) (2020-01-14)
+
+
+### Bug Fixes
+
+* issue with document highlighting offset ([#46](https://github.com/watson-developer-cloud/discovery-components/issues/46)) ([9eab866](https://github.com/watson-developer-cloud/discovery-components/commit/9eab866))
+* make DocPrev only display passage text with highlighting ([#50](https://github.com/watson-developer-cloud/discovery-components/issues/50)) ([51fb55a](https://github.com/watson-developer-cloud/discovery-components/commit/51fb55a))
+* pr to add PR13 into janFixpack branch ([#47](https://github.com/watson-developer-cloud/discovery-components/issues/47)) ([bbdb883](https://github.com/watson-developer-cloud/discovery-components/commit/bbdb883))
 
 
 
@@ -116,7 +207,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 * structured query add and remove rules ([#27](https://github.com/watson-developer-cloud/discovery-components/issues/27)) ([eab1d3b](https://github.com/watson-developer-cloud/discovery-components/commit/eab1d3b))
 * structured query basic ([#22](https://github.com/watson-developer-cloud/discovery-components/issues/22)) ([b992f63](https://github.com/watson-developer-cloud/discovery-components/commit/b992f63))
 * upgrade storybook to 5.3 ([#25](https://github.com/watson-developer-cloud/discovery-components/issues/25)) ([93db599](https://github.com/watson-developer-cloud/discovery-components/commit/93db599))
-